Move creation logic to own implementation.
